app/requirements.txt
flask
flask-sqlalchemy
flask-login
mysql-connector-python